Move develop forking logic into gatsby/commands/develop
