+++ packages/gatsby-cli/package.json
@@ -16,7 +16,6 @@
     "better-opn": "^1.0.0",
     "bluebird": "^3.7.2",
     "chalk": "^2.4.2",
-    "chokidar": "^3.3.1",
     "clipboardy": "^2.2.0",
     "common-tags": "^1.8.0",
     "configstore": "^5.0.1",
@@ -41,7 +40,6 @@
     "react": "^16.8.0",
     "redux": "^4.0.5",
     "resolve-cwd": "^2.0.0",
-    "respawn": "^2.6.0",
     "semver": "^6.3.0",
     "signal-exit": "^3.0.2",
     "source-map": "0.7.3",
